backend/requirements.txt
Flask==2.2.5
python-dotenv==1.0.1
langchain-community
langchain-google-genai
langchain-text-splitters
langchain-groq>=0.1.0  # Groq LLM integration
langchain-pinecone>=0.0.5  # More recent version
pinecone-client>=3,<4
SpeechRecognition==3.10.0
moviepy>=1.0.3
Werkzeug>=2.2.2  # Updated to remove the conflict
flask_cors
pypdf
pdf2image
pytesseract
scikit-learn
gunicorn
